pyproject.toml
[tool.poetry]
name = "gamepad_example"
version = "0.0.1"
description = ""
readme = "README.md"
authors = [
    "Wandelbots <info@wandelbots.com>",
    "Florian Jung <florian.jung@wandelbots.com>",
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
loguru = "^0.7.2"
wandelbots-api-client = "24.1.34"
python-decouple = "^3.8"
requests = "^2.32.3"
inputs = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint = "2.17"

[tool.poetry.scripts]
serve = "gamepad_example:main"
